# openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "API Minimarket",
                 "description":  "DocumentaciÃ³n de la API del sistema Minimarket",
                 "version":  "1.0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Generated server url"
                    }
                ],
...
    "paths":  {
...
                  "/api/carrito/usuario/{usuarioId}/agregar":  {
                                                                   "post":  {
                                                                                "tags":  [
                                                                                             "Carrito"
                                                                                         ],
                                                                                "summary":  "Agregar un nuevo producto al carrito",
                                                                                "description":  "Agrega un nuevo producto al carrito del usuario y devuelve los detalles del carrito actualizado.",
                                                                                "operationId":  "agregarProducto",
                                                                                "parameters":  [
                                                                                                   {
                                                                                                       "name":  "usuarioId",
                                                                                                       "in":  "path",
                                                                                                       "required":  true,
                                                                                                       "schema":  {
                                                                                                                      "type":  "integer",
                                                                                                                      "format":  "int64"
                                                                                                                  }
                                                                                                   },
                                                                                                   {
                                                                                                       "name":  "productoId",
                                                                                                       "in":  "query",
                                                                                                       "required":  true,
                                                                                                       "schema":  {
                                                                                                                      "type":  "integer",
                                                                                                                      "format":  "int64"
                                                                                                                  }
                                                                                                   },
                                                                                                   {
                                                                                                       "name":  "cantidad",
                                                                                                       "in":  "query",
                                                                                                       "required":  true,
                                                                                                       "schema":  {
                                                                                                                      "type":  "integer",
                                                                                                                      "format":  "int32"
                                                                                                                  }
                                                                                                   }
                                                                                               ],
                                                                                "responses":  {
                                                                                                  "201":  {
                                                                                                              "description":  "Producto agregado al carrito exitosamente",
                                                                                                              "content":  {
                                                                                                                              "*/*":  {
                                                                                                                                          "schema":  {
                                                                                                                                                         "$ref":  "#/components/schemas/EntityModelCarrito"
                                                                                                                                                     }
                                                                                                                                      }
                                                                                                                          }
                                                                                                          }
                                                                                              },
                                                                                "security":  [
                                                                                                 {
                                                                                                     "bearerAuth":  [

                                                                                                                    ]
                                                                                                 }
                                                                                             ]
                                                                            }
                                                               },
...
                  "/api/carrito/usuario/{usuarioId}":  {
                                                           "get":  {
                                                                       "tags":  [
                                                                                    "Carrito"
                                                                                ],
                                                                       "summary":  "Obtener carrito por ID de usuario",
                                                                       "description":  "Devuelve los detalles de un carrito especÃ­fico segÃºn el ID del usuario.",
                                                                       "operationId":  "obtenerCarrito",
                                                                       "parameters":  [
                                                                                          {
                                                                                              "name":  "usuarioId",
                                                                                              "in":  "path",
                                                                                              "required":  true,
                                                                                              "schema":  {
                                                                                                             "type":  "integer",
                                                                                                             "format":  "int64"
                                                                                                         }
                                                                                          }
                                                                                      ],
                                                                       "responses":  {
                                                                                         "200":  {
                                                                                                     "description":  "Carrito obtenido exitosamente",
                                                                                                     "content":  {
                                                                                                                     "*/*":  {
                                                                                                                                 "schema":  {
                                                                                                                                                "$ref":  "#/components/schemas/EntityModelCarrito"
                                                                                                                                            }
                                                                                                                             }
                                                                                                                 }
                                                                                                 }
                                                                                     },
                                                                       "security":  [
                                                                                        {
                                                                                            "bearerAuth":  [

                                                                                                           ]
                                                                                        }
                                                                                    ]
                                                                   }
                                                       },
                  "/api/detalle-carrito/{id}":  {
                                                    "delete":  {
                                                                   "tags":  [
                                                                                "detalle-carrito-controller"
                                                                            ],
                                                                   "operationId":  "eliminarItem",
                                                                   "parameters":  [
                                                                                      {
                                                                                          "name":  "id",
                                                                                          "in":  "path",
                                                                                          "required":  true,
                                                                                          "schema":  {
                                                                                                         "type":  "integer",
                                                                                                         "format":  "int64"
                                                                                                     }
                                                                                      }
                                                                                  ],
                                                                   "responses":  {
                                                                                     "200":  {
                                                                                                 "description":  "OK"
                                                                                             }
                                                                                 }
                                                               }
                                                },
                  "/api/carrito/usuario/{usuarioId}/vaciar":  {
                                                                  "delete":  {
                                                                                 "tags":  [
                                                                                              "Carrito"
                                                                                          ],
                                                                                 "summary":  "Vaciar el carrito",
                                                                                 "description":  "Elimina todos los productos del carrito del usuario.",
                                                                                 "operationId":  "vaciarCarrito",
                                                                                 "parameters":  [
                                                                                                    {
                                                                                                        "name":  "usuarioId",
                                                                                                        "in":  "path",
                                                                                                        "required":  true,
                                                                                                        "schema":  {
                                                                                                                       "type":  "integer",
                                                                                                                       "format":  "int64"
                                                                                                                   }
                                                                                                    }
                                                                                                ],
                                                                                 "responses":  {
                                                                                                   "200":  {
                                                                                                               "description":  "Carrito vaciado exitosamente",
                                                                                                               "content":  {
                                                                                                                               "*/*":  {
                                                                                                                                           "schema":  {
                                                                                                                                                          "type":  "object",
                                                                                                                                                          "additionalProperties":  {
                                                                                                                                                                                       "type":  "string"
                                                                                                                                                                                   }
                                                                                                                                                      }
                                                                                                                                       }
                                                                                                                           }
                                                                                                           }
                                                                                               },
                                                                                 "security":  [
                                                                                                  {
                                                                                                      "bearerAuth":  [

                                                                                                                     ]
                                                                                                  }
                                                                                              ]
                                                                             }
                                                              },
                  "/api/carrito/usuario/{usuarioId}/item/{detalleId}":  {
                                                                            "delete":  {
                                                                                           "tags":  [
                                                                                                        "Carrito"
                                                                                                    ],
                                                                                           "summary":  "Eliminar un producto del carrito",
                                                                                           "description":  "Elimina un producto del carrito del usuario segÃºn su ID.",
                                                                                           "operationId":  "eliminarItem_1",
                                                                                           "parameters":  [
                                                                                                              {
                                                                                                                  "name":  "usuarioId",
                                                                                                                  "in":  "path",
                                                                                                                  "required":  true,
                                                                                                                  "schema":  {
                                                                                                                                 "type":  "integer",
                                                                                                                                 "format":  "int64"
                                                                                                                             }
                                                                                                              },
                                                                                                              {
                                                                                                                  "name":  "detalleId",
                                                                                                                  "in":  "path",
                                                                                                                  "required":  true,
                                                                                                                  "schema":  {
                                                                                                                                 "type":  "integer",
                                                                                                                                 "format":  "int64"
                                                                                                                             }
                                                                                                              }
                                                                                                          ],
                                                                                           "responses":  {
                                                                                                             "200":  {
                                                                                                                         "description":  "Producto eliminado exitosamente",
                                                                                                                         "content":  {
                                                                                                                                         "*/*":  {
                                                                                                                                                     "schema":  {
                                                                                                                                                                    "$ref":  "#/components/schemas/EntityModelCarrito"
                                                                                                                                                                }
                                                                                                                                                 }
                                                                                                                                     }
                                                                                                                     }
                                                                                                         },
                                                                                           "security":  [
                                                                                                            {
                                                                                                                "bearerAuth":  [

                                                                                                                               ]
                                                                                                            }
                                                                                                        ]
                                                                                       }
                                                                        }
              },
    "components":  {
                       "schemas":  {
...
                                       "DetalleCarrito":  {
                                                              "type":  "object",
                                                              "properties":  {
                                                                                 "id":  {
                                                                                            "type":  "integer",
                                                                                            "format":  "int64"
                                                                                        },
                                                                                 "producto":  {
                                                                                                  "$ref":  "#/components/schemas/Producto"
                                                                                              },
                                                                                 "cantidad":  {
                                                                                                  "type":  "integer",
                                                                                                  "format":  "int32"
                                                                                              }
                                                                             }
                                                          },
...
                                       "EntityModelCarrito":  {
                                                                  "type":  "object",
                                                                  "properties":  {
                                                                                     "id":  {
                                                                                                "type":  "integer",
                                                                                                "format":  "int64"
                                                                                            },
                                                                                     "items":  {
                                                                                                   "type":  "array",
                                                                                                   "items":  {
                                                                                                                 "$ref":  "#/components/schemas/DetalleCarrito"
                                                                                                             }
                                                                                               },
                                                                                     "_links":  {
                                                                                                    "$ref":  "#/components/schemas/Links"
                                                                                                }
                                                                                 }
                                                              },
...
                                       "Links":  {
                                                     "type":  "object",
                                                     "additionalProperties":  {
                                                                                  "$ref":  "#/components/schemas/Link"
                                                                              }
                                                 }
                                   },
                       "securitySchemes":  {
                                               "bearerAuth":  {
                                                                  "type":  "http",
                                                                  "description":  "AutenticaciÃ³n JWT usando Bearer Token",
                                                                  "in":  "header",
                                                                  "scheme":  "bearer",
                                                                  "bearerFormat":  "JWT"
                                                              }
                                           }
                   }
}